% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brenton2019}
\alias{brenton2019}
\title{Data from Benton et al. (2019)}
\format{A data frame with 881 rows and 3 variables:
\describe{
  \item{A}{APAS values; \code{1}=not significan growth (0 and 10^6 CFU/L), \code{2}=significant growth (10^7 and 10^8+ CFU/L)}
  \item{S}{St Vincent's standard workflow values; \code{1}=not significan growth (0 and 10^6 CFU/L), \code{2}=significant growth (10^7 and 10^8+ CFU/L)}
  \item{P}{Panel consensus values; \code{1}=not significan growth (0 and 10^6 CFU/L), \code{2}=significant growth (10^7 and 10^8+ CFU/L)}
}}
\usage{
data(brenton2019)
}
\description{
These are the data from from Benton et al. (2019) containing the APAS (A, the index test),
St Vincent's standard workflow (S, imperfect truth) and the panel consensus (P, resolver) variables.
}
\keyword{datasets}
