% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hawkins2001}
\alias{hawkins2001}
\title{Representation of the data from Hawkins et al. (2001)}
\format{A data frame with 3,000 rows and 3 variables:
\describe{
  \item{index}{index test values; \code{1}=negative, \code{2}=positive}
  \item{ref}{Reference values; \code{1}=negative, \code{2}=positive}
  \item{resolve}{Resolver values; \code{1}=negative, \code{2}=positive}
}}
\usage{
data(hawkins2001)
}
\description{
These are a recreation of a sample line data that prodice the cross
tabulations seen in Hawkins et al. (2001) containing the index test (\code{index}),
reference (\code{ref}) and the resolver (\code{resolve}) variables.
}
\keyword{datasets}
