% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_crs.R
\name{perform_crs}
\alias{perform_crs}
\title{Perform CRS sens and spec calulations using a one sample-per-row input dataset}
\usage{
perform_crs(dat, index, imperfect, resolver, trans_method = "probit",
  alpha = 0.05)
}
\arguments{
\item{dat}{data frame that contains the index test, imperfect truth and resolver columns (see details)}

\item{index}{string of column name in \code{dat} corresponding to the index test (1=negative, 2=positive)}

\item{imperfect}{string of column name in \code{dat} corresponding to the imperfect truth (1=negative, 2=positive)}

\item{resolver}{string of column name in \code{dat} corresponding to the resolver (1=negative, 2=positive, NA=not known)}

\item{trans_method}{one of "probit" (default), "logit", "loglog", "arcsin"}

\item{alpha}{confidence level (default=0.05)}
}
\value{
A tibble with columns \code{param}, \code{p}, \code{var_p}, \code{se_p}, \code{p_lo}, \code{p_up}
}
\description{
Performs the CRS analysis
}
\details{
Currently \code{perform_mcrs} requires \code{dat} to contain variables with negative and positive values as
    the integers \code{1}=negative and \code{2}=positive
}
\examples{
library(tibble)
data(brenton2019)
perform_crs(brenton2019, "A", "S", "P")
}
\author{
Ty Stanford <tyman@lbtinnovations.com>
}
